% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.scampr.R
\name{AIC.scampr}
\alias{AIC.scampr}
\title{Akaike's Information Criteria for objects of class 'scampr'}
\usage{
\method{AIC}{scampr}(object, k = 2)
}
\arguments{
\item{object}{a scampr model}
}
\value{

}
\description{
Akaike's Information Criteria for objects of class 'scampr'
}
