% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.scampr.R
\name{BIC.scampr}
\alias{BIC.scampr}
\title{Wrapper for Akaike's Information Criteria for objects of class 'scampr' with k = log(N)}
\usage{
\method{BIC}{scampr}(object, ...)
}
\arguments{
\item{object}{a scampr model}

\item{...}{Optionally, additional scampr model objects}
}
\value{
a numeric value with the corresponding BIC
}
\description{
Wrapper for Akaike's Information Criteria for objects of class 'scampr' with k = log(N)
}
\examples{
# Get the gorilla nesting data
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Fit a scampr model to the point pattern
m <- scampr(pres ~ elev.std, data = dat, model.type = "ipp")

BIC(m)
}
