% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{bound}
\alias{bound}
\title{Calculate the approximate boundary of 2D Euclidean data}
\usage{
bound(
  object,
  coord.names = c("x", "y"),
  precision = 1,
  plotting = F,
  encase = NA,
  smoothed = F
)
}
\arguments{
\item{object}{Either a data frame containing the spatial locations of each point or a \code{scampr} model object. In the latter case the model's data will be used.}

\item{coord.names}{a vector of length 2 containing character strings describing the column names of the coordinates in any data provided. First coordinate name should refer to the horizontal axis. Will be over-ridden when \code{object} is a \code{scampr} model.}

\item{precision}{an integer value that determines the number of decimal places to search through. Controls the precision of the approximate boundary, 1 or 2 should be adequate. The higher (larger) the precision the slower the result is obtained.}

\item{plotting}{a logical indicating whether the results should be plotted. Helpful for trialing presicion and/or encasing.}

\item{encase}{a numeric value indicating the amount of encasing of the data should occur. This will essentially add a buffer to the boundary.}

\item{smoothed}{a logical indicating whether or not the boundary should be smoothed using a Loess' smoother.}
}
\value{
a data frame of 2 columns for each coordinate.
}
\description{
Calculate the approximate boundary of 2D Euclidean data
}
\examples{
# Get the gorilla nesting data
dat <- gorillas

# Calculate a boundary for the data
bnd <- bound(dat, coord.names = c("x", "y"))
}
