% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_test.R
\name{cluster_test}
\alias{cluster_test}
\title{Perform simulations from a scampr model to calculate global envelopes for the inhomogeneous K function}
\usage{
cluster_test(
  model,
  conf.level = 0.95,
  nsims = 100,
  Kfun.correction = c("border", "bord.modif", "isotropic", "translate"),
  return.data = F
)
}
\arguments{
\item{model}{a scampr model object}

\item{conf.level}{a numeric between 0 and 1 indicating the confidence level of the global envelope test}

\item{nsims}{an integer describing the number of simulations to conduct}

\item{Kfun.correction}{a character string describing the correction type to be used. See \code{spatstat::Kinhom()} for details.}

\item{return.data}{a logical indicating whether to return the observed and simulated data}
}
\value{
Optionally, a list of the observed and simulated K functions - see \code{GET::create_curve_set()}. Results are plotted.
}
\description{
Perform simulations from a scampr model to calculate global envelopes for the inhomogeneous K function
}
\examples{
# Get the data
dat <- scampr::gorillas
dat$elev <- scale(dat$elevation)
mod <- po(pres ~ elev, dat, model.type = "ipp")
cluster_test(mod)
}
