% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.scampr.R
\name{coef.scampr}
\alias{coef.scampr}
\title{Coefficients for objects of class 'scampr'}
\usage{
coef.scampr(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Coefficients for objects of class 'scampr'
}
