% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.scampr.R
\name{coefficients.scampr}
\alias{coefficients.scampr}
\title{Coefficients for objects of class 'scampr'}
\usage{
\method{coefficients}{scampr}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Coefficients for objects of class 'scampr'
}
