% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.bf.matrix.R
\name{get.bf.matrix}
\alias{get.bf.matrix}
\title{Internal scampr function that calculates the basis function evaluations at point locations provided. From either a fitted model or data.frame of class 'bf.df'}
\usage{
get.bf.matrix(object, point.locations)
}
\arguments{
\item{object}{either a fitted model or data.frame of class 'bf.df' created by 'simple_basis'}

\item{point.locations}{either a matrix or data.frame describing the point locations in the same order as in 'object'}
}
\value{

}
\description{
Internal scampr function that calculates the basis function evaluations at point locations provided. From either a fitted model or data.frame of class 'bf.df'
}
