% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.data.R
\name{get.data}
\alias{get.data}
\title{Internal scampr function that extracts data list of presence and quadrature from a fitted model}
\usage{
get.data(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Internal scampr function that extracts data list of presence and quadrature from a fitted model
}
