% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.design.matrix.R
\name{get.desgin.matrix}
\alias{get.desgin.matrix}
\title{Internal scampr function that creates a fixed effect design matrix from a formula and data}
\usage{
get.desgin.matrix(formula, data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Internal scampr function that creates a fixed effect design matrix from a formula and data
}
