% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.offset.R
\name{get.offset}
\alias{get.offset}
\title{Internal scampr function that creates a fixed effect offset vector from a formula and data}
\usage{
get.offset(formula, data, subset, na.action)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of a model. See GLM function for further formula details.}

\item{data}{a data frame containing predictors and response of the formula parameter.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another possible value is NULL, no action. Value na.exclude can be useful.}
}
\value{
a data.frame
}
\description{
Internal scampr function that creates a fixed effect offset vector from a formula and data
}
\examples{
des.mat <- scampr:::get.design.matrix(Petal.Length ~ Petal.Width + Species, iris)
head(des.mat)
}
