% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.TMB.data.input.R
\name{get.TMB.data.input}
\alias{get.TMB.data.input}
\title{Internal scampr function that creates a list of data and starting parameters for scampr models}
\usage{
get.TMB.data.input(
  po.formula,
  pa.formula,
  po.data,
  pa.data,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  FRK.basis.functions,
  simple.basis,
  model.type = c("laplace", "variational", "ipp"),
  bf.matrix.type = c("sparse", "dense"),
  data.type = c("po", "pa", "popa"),
  starting.pars
)
}
\arguments{
\item{starting.pars}{}
}
\value{

}
\description{
Internal scampr function that creates a list of data and starting parameters for scampr models
}
