% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCall.scampr.R
\name{getCall.scampr}
\alias{getCall.scampr}
\title{getCall function for objects of class 'scampr'}
\usage{
\method{getCall}{scampr}(x, ...)
}
\arguments{
\item{x}{a scampr model object}

\item{...}{Additional arguments to the call, or arguments with changed values}
}
\value{
the updated call
}
\description{
get the call from a scampr model
}
\examples{
# Get the flora PO data for one of the species
dat_po <- flora$po$sp1

# Attach the quadrature to the PO data
dat_po <- rbind.data.frame(dat_po, flora$quad)

# Fit an IPP model to the point pattern
m.ipp <- scampr(pres ~ MNT + D.Main, dat_po, model.type = "ipp")

getCall(m.ipp)
}
