% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.scampr.R
\name{image.scampr}
\alias{image.scampr}
\title{Plot the image of a field over a model quadrature}
\usage{
\method{image}{scampr}(x, z, residual.type, ...)
}
\arguments{
\item{x}{A scampr model object}

\item{z}{Either a single characater string of the variable name (in the model data) or one of 'fitted', 'residuals'. Alternatively, a vector of numeric values to be plotted}

\item{residual.type}{an optional character string for residual type if z == 'residual'}

\item{...}{}
}
\value{

}
\description{
Plot the image of a field over a model quadrature
}
