% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipp.R
\name{ipp}
\alias{ipp}
\title{Inhomogeneous Poisson Process Model for Point Patterns}
\usage{
ipp(
  formula,
  data,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  starting.pars,
  se = TRUE,
  subset
)
}
\arguments{
\item{formula}{formula describing fixed effects of the linear predictor. Response must be the presence/quadrature identifier.}

\item{data}{data frame containing predictors at both presence-records and quadrature}

\item{coord.names}{vector of character strings describing the column names of the coordinates in data}

\item{quad.weights.name}{charater string of the column name of quadrature weights in data}

\item{starting.pars}{optional named list or scampr model that gives warm starts for the parameters of the model}

\item{se}{logical indicating whether standard errors should be calculated}

\item{subset}{an optional vector describing a subset of the data to be used.}
}
\value{
scampr model object
}
\description{
Fit an inhomogeneous Poisson process model to a point pattern using numerical quadrature (provided with the data, see e.g. scampr::gorillas)
}
\examples{
# Get the gorilla nesting data
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Fit an IPP model to the point pattern
m.ipp <- ipp(pres ~ elev.std, data = dat)
}
