% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ippm.R
\name{ippm}
\alias{ippm}
\title{Inhomogeneous Poisson Process}
\usage{
ippm(formula, data, starting.pars = NULL, se = TRUE)
}
\arguments{
\item{formula}{formula describing fixed effects of the linear predictor. Response must be the presence/quadrature identifier.}

\item{data}{data frame containing predictors at both presence-records and quadrature}

\item{starting.pars}{optional named list or scampr model that gives warm starts for the parameters of the model}

\item{se}{logical indicating whether standard errors should be calculated}
}
\value{
scampr model object
}
\description{
Fit an inhomogeneous Poisson process model to presence records using numerical quadrature
}
