% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfunc.R
\name{kfunc}
\alias{kfunc}
\title{Inhomogeneous K function for scampr models (is a wrapper of \code{spatstat}'s version). Or calculate the K function from point pattern (and intensity) provided.}
\usage{
kfunc(
  model,
  correction = c("border", "bord.modif", "isotropic", "translate"),
  point.pattern,
  intensity.at.pp,
  spatstat.win,
  intensity.at = c("im", "pts"),
  dists = NULL
)
}
\arguments{
\item{model}{a scampr model object}

\item{correction}{a character string describing the correction type to be used. See \code{spatstat.explore::Kinhom()} for details.}

\item{point.pattern}{Optionally, a data frame, the first two columns of which describe the horizontal and vertical coordinates of point locations respectively.}

\item{intensity.at.pp}{Optionally, a vector of length \code{nrow(point.pattern)} that describes the intensity at each point.}

\item{spatstat.win}{Optionally, a spatstat window object.}

\item{intensity.at}{Optionally, a character string, one of either 'im' or 'pts'. Describes whether the intensity used in \code{spatstat.explore::Kinhom()} is an image object or just provided at the presence points.}

\item{dists}{a vector of values for the distances at which the inhomogeneous K function should be evaluated. Not normally given by the user; \code{spatstat} provides a sensible default.}
}
\value{
a data.frame of two columns - distance (dist) and corresponding K function values (Kfn).
}
\description{
Inhomogeneous K function for scampr models (is a wrapper of \code{spatstat}'s version). Or calculate the K function from point pattern (and intensity) provided.
}
\examples{
# Get the data
dat <- scampr::gorillas
dat$elev <- scale(dat$elevation)
mod <- scampr(pres ~ elev, dat, model.type = "ipp")
K_func <- kfunc(mod)
}
