% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfunc_envelopes.R
\name{kfunc_envelopes}
\alias{kfunc_envelopes}
\title{Perform simulations from a scampr model to calculate global envelopes for the inhomogeneous K function}
\usage{
kfunc_envelopes(
  model,
  conf.level = 0.95,
  nsims = 100,
  Kfun.correction = c("border", "bord.modif", "isotropic", "translate"),
  return.data = F,
  spatstat.win,
  dists = NULL
)
}
\arguments{
\item{model}{a scampr model object}

\item{conf.level}{a numeric between 0 and 1 indicating the confidence level of the global envelope test}

\item{nsims}{an integer describing the number of simulations to conduct}

\item{Kfun.correction}{a character string describing the correction type to be used. See \code{spatstat::Kinhom()} for details.}

\item{return.data}{a logical indicating whether to return the observed and simulated data}

\item{spatstat.win}{Optionally, a spatstat window object.}

\item{dists}{a vector of values for the distances at which the inhomogeneous K function should be evaluated. Not normally given by the user; \code{spatstat} provides a sensible default.}
}
\value{
Optionally, a list of the observed and simulated K functions - see \code{GET::create_curve_set()}. Results are plotted.
}
\description{
Perform simulations from a scampr model to calculate global envelopes for the inhomogeneous K function
}
\examples{
# Get the data
dat <- scampr::gorillas
dat$elev <- scale(dat$elevation)
mod <- scampr(pres ~ elev, dat, model.type = "PO", include.sre = F)
kfunc_envelopes(mod)
}
