% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcp.R
\name{lgcp}
\alias{lgcp}
\title{Approximate log-Gaussian Cox Process Models for Point Patterns}
\usage{
lgcp(
  formula,
  data,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  FRK.basis.functions,
  simple.basis,
  approx.with = c("laplace", "variational"),
  se = TRUE,
  bf.matrix.type = c("sparse", "dense"),
  starting.pars,
  subset
)
}
\arguments{
\item{formula}{formula describing fixed effects of the linear predictor. Response must be the presence/quadrature binary (1/0) column name.}

\item{data}{data frame containing predictors at both presence-records and quadrature}

\item{coord.names}{vector of character strings describing the column names of the coordinates in data}

\item{quad.weights.name}{charater string of the column name of quadrature weights in data}

\item{FRK.basis.functions}{object of class 'Basis' from FRK package. If neither 'FRK.basis.functions' nor 'simple.basis' is specified will use default FRK::auto_basis with 2 spatial resolutions}

\item{simple.basis}{Alternative to 'FRK.basis.functions', a data.frame of basis functions information created by 'simple_basis()'.}

\item{approx.with}{charater indicating the type of approximation to use for the intractible marginalisation: variational or Laplace}

\item{se}{logical indicating whether standard errors should be calculated}

\item{bf.matrix.type}{a character string, one of 'sparse' or 'dense' indicating whether to use sparse or dense matrix computations for the basis functions created.}

\item{starting.pars}{optional named list or scampr model that gives warm starts for the parameters of the model}

\item{subset}{an optional vector describing a subset of the data to be used.}
}
\value{
a scampr model object.
}
\description{
Fit a log-Gaussian Cox process model to a point pattern using numerical quadrature (provided with the data, see e.g. scampr::gorillas) and one of either a Laplace or variational approximation to marginalise over the latent field.
}
\examples{
#' # Get the gorilla nesting data
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Set up a simple 2D grid of basis functions to fit a LGCP model to the data
bfs <- simple_basis(nodes.on.long.edge = 9, data = dat)

# Fit a LGCP model using variational approximation
m.lgcp_va <- lgcp(pres ~ elev.std, data = dat, approx.with = "variational", simple.basis = bfs)

\dontrun{
# Fit a LGCP model using Laplace approximation
m.lgcp_lp <- lgcp(pres ~ elev.std, data = dat, approx.with = "laplace", simple.basis = bfs)
}
}
