% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpm.R
\name{lgcpm}
\alias{lgcpm}
\title{Approximate log-Gaussian Cox Process Model for a Point Pattern}
\usage{
lgcpm(
  formula,
  data,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  FRK.basis.functions,
  simple.basis,
  approx.with = c("laplace", "variational"),
  se = TRUE,
  bf.matrix.type = c("sparse", "dense"),
  starting.pars,
  subset,
  na.action
)
}
\arguments{
\item{formula}{formula describing fixed effects of the linear predictor. Response must be the presence/quadrature binary (1/0) column name.}

\item{data}{data frame containing predictors at both presence-records and quadrature}

\item{coord.names}{vector of character strings describing the column names of the coordinates in data}

\item{quad.weights.name}{charater string of the column name of quadrature weights in data}

\item{FRK.basis.functions}{object of class 'Basis' from FRK package. If neither 'FRK.basis.functions' nor 'simple.basis' is specified will use default FRK::auto_basis with 2 spatial resolutions}

\item{simple.basis}{Alternative to 'FRK.basis.functions', a data.frame of basis functions information created by 'simple_basis()'.}

\item{approx.with}{charater indicating the type of approximation to use for the intractible marginalisation: variational or Laplace}

\item{se}{logical indicating whether standard errors should be calculated}

\item{bf.matrix.type}{a character string, one of 'sparse' or 'dense' indicating whether to use sparse or dense matrix computations for the basis functions created.}

\item{starting.pars}{optional named list or scampr model that gives warm starts for the parameters of the model}

\item{subset}{optional subset of the data}

\item{na.action}{optional way of handling NA's in the data, default is omit}
}
\value{
a scampr model object
}
\description{
Blah blah blah
}
\examples{
#' # Get the gorilla nesting data
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Set up a simple 2D grid of basis functions to fit a LGCP model to the data
bfs <- simple_basis(nodes.on.long.edge = 9, data = dat)

# Fit a LGCP model using variational approximation
m.lgcp_va <- lgcpm(pres ~ elev.std, data = dat, approx.with = "variational", simple.basis = bfs)

#' # Fit a LGCP model using Laplace approximation
m.lgcp_lp <- lgcpm(pres ~ elev.std, data = dat, approx.with = "laplace", simple.basis = bfs)
}
