% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.scampr.R
\name{logLik.scampr}
\alias{logLik.scampr}
\title{Marginal log-likelihood for objects of class 'scampr'}
\usage{
\method{logLik}{scampr}(object)
}
\arguments{
\item{object}{}
}
\value{
numeric describing the approximate marginal log-likelihood of the scampr model object
}
\description{
Marginal log-likelihood for objects of class 'scampr'
}
