% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.spatial.folds.R
\name{make.spatial.folds}
\alias{make.spatial.folds}
\title{Partition data into k spatially blocked folds for cross validation}
\usage{
make.spatial.folds(data, rangeX, rangeY, k = 5, coord.names = c("x", "y"))
}
\arguments{
\item{data}{a data frame with columns \code{coord.names} to be split into spatial k folds.}

\item{rangeX}{a vector of length 2 denoting the range of the horizontal coordinate. If missing this will be obtained from \code{data} - note that this assumes the region of interest is completely represented by points in the data frame provided.}

\item{rangeY}{a vector of length 2 denoting the range of the vertical coordinate. If missing this will be obtained from \code{data} - note that this assumes the region of interest is completely represented by points in the data frame provided.}

\item{k}{an integer denoting the number of folds. Default is 5.}

\item{coord.names}{a vector of length two describing the column names of the horizontal and vertical coordinates in \code{data} in this order. Default is "x" and "y" respectively.}
}
\value{
a vector of length \code{nrow(data)} with integers denoting the fold into which each point falls.
}
\description{
This function takes in a data frame containing coordinates and creates a spatially blocked set of k folds.
}
\examples{
#' # Get the gorilla nesting data
dat <- gorillas

# create spatial CV folds
dat$fold <- make.spatial.folds(dat)

# view the folds
table(dat$fold)
}
