% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FRK_bf_matrix.R
\name{make_FRK_bf_matrix}
\alias{make_FRK_bf_matrix}
\title{Create an FRK basis function matrix}
\usage{
make_FRK_bf_matrix(
  data,
  coord.names = c("x", "y"),
  manifold = FRK::plane(),
  regular = 1,
  nres = 2,
  prune = 0,
  max_basis = NULL,
  subsamp = 10000,
  type = c("bisquare", "Gaussian", "exp", "Matern32"),
  isea3h_lo = 2,
  bndary = NULL,
  scale_aperture = ifelse(is(manifold, "sphere"), 1, 1.25),
  verbose = 0L,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create an FRK basis function matrix
}
