% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scampr.R
\name{plot.scampr}
\alias{plot.scampr}
\title{Plot diagnositcs for objects of class 'scampr'}
\usage{
\method{plot}{scampr}(x, ..., which = c("residuals", "fitted"), add.points = F)
}
\arguments{
\item{x}{a scampr model object}

\item{...}{additional plotting arguments}

\item{which}{if a subset of the plots is required, specify one of 'residuals' or 'fitted': plots Pearson residuals and fitted intensity respectively for PPM (for a PA data model: plots fitted vs. raw residuals and the survey data for 'residuals' and 'fitted' resp.).}

\item{add.points}{logical indicating whether to add the presence point locations to plots}
}
\value{
See spatstat::plot.ppm
}
\description{
Currently plots residuals as in 'spatstat', as well as the fitted intensity, over the quadrature used to model the point pattern. In the case of presence/absence data fitted vs. residuals is plotted.
}
\examples{
# Get the Eucalypt data
dat_po <- eucalypt[["po"]]
dat_pa <- eucalypt[["pa"]]

# Fit an IPP model to the point pattern
m.ipp <- ipp(pres ~ TMP_MIN + D_MAIN_RDS, data = dat_po)

# Fit a combined data model
m.popa <- popa(pres ~ TMP_MIN + D_MAIN_RDS, Y ~ TMP_MIN,
po.data = dat_po, pa.data = dat_pa, model.type = "ipp")

# Fit presence/absence model
m.pa <- pa(Y ~ TMP_MIN, pa.data = dat_pa, model.type = "ipp")

plot(m.ipp)
plot(m.popa)
plot(m.pa)
}
