% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basis.R
\name{plot_basis}
\alias{plot_basis}
\title{Plot basis functions for objects of class 'bf.df' ### WIP ###}
\usage{
plot_basis(x, ..., add = FALSE)
}
\arguments{
\item{x}{a basis function data frame ("bf.df") object}

\item{...}{additional plotting arguments}

\item{add}{logical indicating whether to add the basis functions to the current plot}
}
\description{
Plots the bi-square basis functions created by \code{scampr::simple_basis()}.
}
\examples{
# Create the basis functions on the flora quadrature
bfs <-  simple_basis(9, flora$quad)

# Plot the bi-square basis functions
plot_basis(bfs)
}
