% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_field.R
\name{plot_field}
\alias{plot_field}
\title{Plot the 2D fields from a scampr model ### WIP ###}
\usage{
plot_field(
  object,
  domain.data,
  field = c("log Lambda", "Presence Probability", "Fixed Effects",
    "Fixed Bias Effects", "Latent Field", "Latent Bias Field"),
  ...
)
}
\arguments{
\item{object}{a scampr model object}

\item{domain.data}{Optional. A data frame of columns 'coord.names' that contains at least the extremities of the domain of interest. If missing the quadrature points from the model will be used - Note that if these do not form a cover of the domain the image will not work (likewise, will not work for scampr models that do not include quadrature).}

\item{field}{}

\item{...}{additional plotting arguments}
}
\value{
See spatstat::plot.im
}
\description{
Plots a selected field from a fitted scampr model. Options include the fitted intensity or its components on the log scale.
}
\examples{
# Get the flora data for one of the species
dat_po <- flora$po$sp1
dat_pa <- flora$pa

# obtain a sample of 10,000 quadrature points for the point process model
set.seed(1)
quad.pts <- flora$quad[sample(1:nrow(flora$quad), 10000, replace = F), ]
set.seed(NULL)

# Attach the quadrature points to the presence-only data
dat_po <- rbind.data.frame(dat_po, quad.pts)

# Create the basis functions on the flora quadrature
bfs <-  simple_basis(9, flora$quad)

# Combined Data Model
m.comb <- scampr(pres ~ MNT, dat_po, bias.formula = ~ D.Main, dat_pa, include.sre = F)

# Plot the fitted log-intensity
plot_field(m.comb, domain.data = flora$quad)
}
