% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popa.R
\name{popa}
\alias{popa}
\title{Combined Model for Presence-only and Presence-Absence Data}
\usage{
popa(
  po.formula,
  pa.formula,
  po.data,
  pa.data,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  FRK.basis.functions,
  simple.basis,
  model.type = c("laplace", "variational", "ipp"),
  bf.matrix.type = c("sparse", "dense"),
  se = TRUE,
  starting.pars,
  subset,
  na.action
)
}
\arguments{
\item{po.formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the Presence-only data model to be fitted. The 'response' must be a binary that indicates whether a datum is a presence or quadrature point. See GLM function for further formula details.}

\item{pa.formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the Presence/Absence data model to be fitted. The 'response' must be a must be either the site abundance or a binary indicating whether there is a presence or not. See GLM function for further formula details.}

\item{po.data}{a data frame containing predictors at both presence-records and quadrature as well as the po.formula 'response'.}

\item{pa.data}{a data frame containing predictors and response for the pa.formula.}

\item{coord.names}{a vector of character strings describing the column names of the coordinates in both data frames}

\item{quad.weights.name}{a charater string of the column name of quadrature weights in the po.data}

\item{FRK.basis.functions}{an optional object of class 'Basis' from FRK package. If neither 'FRK.basis.functions' nor 'simple.basis' is specified will use default FRK::auto_basis with 2 spatial resolutions}

\item{simple.basis}{an alternative to 'FRK.basis.functions': a data.frame of basis functions information created by 'simple_basis()'.}

\item{model.type}{a character string indicating the type of model to be used. May be one of 'laplace' or 'variational' for Cox Processes involving spatially correlated errors or 'ipp' for a model that follows an inhomgeneous Poisson process.}

\item{bf.matrix.type}{a character string, one of 'sparse' or 'dense' indicating whether to use sparse or dense matrix computations for the basis functions created.}

\item{se}{a logical indicating whether standard errors should be calculated}

\item{starting.pars}{an optional named list or scampr model that gives warm starting values for the parameters of the model.}

\item{subset}{an optional subset of the data to be used.}

\item{na.action}{an optional way of handling NA's in the data, default is omit.}
}
\value{

}
\description{
Combined Model for Presence-only and Presence-Absence Data
}
