% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.scampr.R
\name{predict.scampr}
\alias{predict.scampr}
\title{Predict function for objects of class 'scampr'}
\usage{
predict.scampr(
  object,
  newdata,
  type = c("link", "response"),
  dens = c("posterior", "prior"),
  process = c("intensity", "abundance")
)
}
\arguments{
\item{object}{a scampr model object}

\item{newdata}{a data frame of point locations to predict over as well as predictors involved in the model}

\item{type}{a character string indicating the type of linear predictor to be returned. One of 'link' or 'response'}

\item{dens}{a character string indicating the pdf of the random effects to draw from}

\item{process}{a character string indictating the process to be estimated. one of 'intensity' or 'abundance'. Only available for combined data models.}
}
\value{

}
\description{
Predict function for objects of class 'scampr'
}
