% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_simple_basis.R
\name{prune_simple_basis}
\alias{prune_simple_basis}
\title{Pruning for regular grids of bi-square basis functions in flat 2D space}
\usage{
prune_simple_basis(pres.locs, simple.bfs, min.points = 1)
}
\arguments{
\item{pres.locs}{a data frame or matrix of 2 columns describing the horizontal and vertical coordinate locations for presence locations on which to base pruning.}

\item{simple.bfs}{a data frame of class \code{bf.df} created by \code{simple_basis()} that describes the unpruned basis functions.}

\item{min.points}{Optional. a postive integer describing the least number of points required within each basis function for it to be retained in the pruning. Default is 1, meaning all basis functions with no points within their radius are pruned.}
}
\value{
a simple basis data frame of class 'bf.df' pruned to the presence locations provided. Consisting of columns: horizontal axis location, vertical axis location, scale (radius), res (resolution id).
}
\description{
Pruning for regular grids of bi-square basis functions in flat 2D space
}
\examples{
#' # Base the basis function nodes on the locations of presence records and quadrature
dat <- dat <- gorillas
bfs <- simple_basis(nodes.on.long.edge = 9, data = dat)
pruned.bfs <- prune_simple_basis(dat[dat$pres == 1, c("x", "y")], bfs)
}
