% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scampr.R
\name{scampr}
\alias{scampr}
\title{Spatially correlated, approximate modelling of presences in R}
\usage{
scampr(
  formula,
  data,
  pa.formula,
  pa.data,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  basis.functions,
  model.type = c("variational", "laplace", "ipp"),
  sparse = TRUE,
  se = TRUE,
  starting.pars,
  subset,
  additional.latent.field = FALSE
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the Presence-only data model to be fitted. The 'response' must be a binary that indicates whether a datum is a presence or quadrature point. See GLM function for further formula details.}

\item{data}{a data frame containing predictors at both point locations and quadrature as well as the formula 'response' found in \code{formula}.}

\item{pa.formula}{Optionally, an object of class "formula" (or one that can be coerced to that class): a symbolic description of the Presence/Absence data model to be fitted. All predictor terms must also be included in the \code{formula} argument. The 'response' must be a must be either the site abundance or a binary indicating whether there is a presence or not. See GLM function for further formula details.}

\item{pa.data}{Optionally, a data frame containing predictors and response described in \code{pa.formula}.}

\item{coord.names}{a vector of length 2 containing character strings describing the column names of the coordinates in any data provided. First coordinate name should refer to the horizontal axis.}

\item{quad.weights.name}{a character string of the column name of quadrature weights in \code{data}.}

\item{basis.functions}{an optional object of class 'Basis' created by \code{FRK::auto_basis()} or 'bf.df' created by \code{scampr::simple_basis()}. Either object describes a set of basis functions for approximating the latent Gaussian field. If NULL the model will use default \code{FRK::auto_basis()} with \code{max_basis = 0.25 * # of points}.}

\item{model.type}{a character string indicating the type of model to be used. May be one of 'laplace' or 'variational' for Cox Processes involving spatially correlated errors or 'ipp' for a model that follows an inhomogeneous Poisson process.}

\item{sparse}{a logical indicating whether sparse matrix calculations should be used. Should only be turned off when using dense basis functions, e.g. Gaussian Kernel with long range parameter.}

\item{se}{a logical indicating whether standard errors should be calculated.}

\item{starting.pars}{an optional named list or previously fit scampr model object that gives warm starting values for the parameters of the model.}

\item{subset}{an optional vector describing a subset of the data to be used.}

\item{additional.latent.field}{a logical indicating whether an additional latent field should be included in the Presence-only data model to capture additional detection bias. Only used in the integrated data model.}
}
\value{
a scampr model object
}
\description{
This is the main function for modelling presences within the \code{scampr} framework. The type of model will depend on the arguments provided. This can be used to model point patterns as a log-Gaussian Cox process (LGCP) or Inhomogeneous Poisson process (IPP), as well as, jointly fitting a model to presence-only and presence/absence data if both are provided. This function can also fit binary regression with a complimentary log-log link function (with optional spatial random effects) when only presence/absence data is provided.

If only \code{formula} and \code{data} are provided, the function will fit either an IPP, or LGCP model to the point pattern (depending on argument \code{model.type}). This uses numerical quadrature (provided with the data, see e.g. scampr::gorillas) to approximate the spatial integral. If fitting a LGCP, uses one of either a Laplace or variational approximation to marginalise over the latent field.

If only \code{pa.formula} and \code{pa.data} are provided, the function will fits a binary regression model to presence/absence data using a complimentary log-log link function. Can accomodate an approximate latent field as spatial random effects (depending on argument \code{model.type}).

If both \code{formula} and \code{pa.formula}, as well as, \code{data} and \code{pa.data} are provided, the function jointly fits a model to presence-only and presence/absence data as linked by response to environmental predictors provided in each formula. The presence-only formula must also contain biasing predictors to account for opportunistic collection. If argument \code{model.type} is not equal to "ipp", then the two data sources will additionally share a latent Gaussian random field.
}
\examples{
# Get the flora data for one of the species
dat_po <- flora$po$sp1
dat_pa <- flora$pa

# Attach the quadrature to the PO data
dat_po <- rbind.data.frame(dat_po, flora$quad)

# Fit models without a latent effects (IPP) #
# Point Process Model
m.ipp <- scampr(pres ~ MNT + D.Main, dat_po, model.type = "ipp")
# Binary Regression
m.bin <- scampr(pa.formula = sp1 ~ MNT, pa.data = dat_pa, model.type = "ipp")
# Combined Data Model
m.comb <- scampr(pres ~ MNT + D.Main, dat_po, sp1 ~ MNT,
dat_pa, model.type = "ipp")

# Set up a simple 2D grid of basis functions to fit a LGCP model to the data
bfs <- simple_basis(nodes.on.long.edge = 9, data = dat_po)

\dontrun{
# Fit with a shared latent field (LGCP) #
# Point Process Model
m.lgcp <- scampr(pres ~ MNT + D.Main, dat_po, basis.functions = bfs)
# Binary Regression with spatial random effects
m.bin_w_sre <- scampr(pa.formula = sp1 ~ MNT, pa.data = dat_pa, basis.functions = bfs)
# Combined Data Model with spatial random effects
m.comb_w_sre <- scampr(pres ~ MNT + D.Main, dat_po, sp1 ~ MNT,
dat_pa, basis.functions = bfs)
}
}
