% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lgcp_pp.R
\name{simulate_lgcp_pp}
\alias{simulate_lgcp_pp}
\title{Simulate a point pattern from log-Gaussian Cox Process}
\usage{
simulate_lgcp_pp(
  Intercept = -3,
  Slope = 1.25,
  latent.variance = 1,
  latent.range = 15,
  rseed = NA
)
}
\arguments{
\item{Intercept}{Fixed effect intercept - used to control the expected number of points}

\item{Slope}{Fixed effect slope - controls the effect of the spatial covariate on the point pattern}

\item{latent.variance}{The marginal variance of the latent Gaussian field}

\item{latent.range}{The correlation range of the latent Gaussian field}

\item{rseed}{integer for setting the random seed}
}
\value{
the simulated data frame with attributes containing sim info
}
\description{
Simulate a point pattern from a LGCP with an intercept and slope on a single deterministic spatial covariate. Over a hectare domain, includes quadrature. Control the correlation range and marginal variance of the latent field.
}
\examples{
# Simulate a point pattern
dat <- sim_lgcp_pp(Intercept = -3, Slope = 1.25, latent.range = 15, rseed = 1)

# Fit an IPP model to the point pattern
m.ipp <- ippm(pres ~ X, data = dat)

# Set up a simple 2D grid of basis functions to fit a LGCP model to the data
bfs <- simple_basis(nodes.on.long.edge = 10, data = dat)

# Fit a LGCP model to the point pattern
m.lgcp_va <- lgcpm(pres ~ X data = dat, approx.with = "variational", simple.basis = bfs)

summary(m.ipp)
summary(m.lgcp_va)
}
