% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model_pp.R
\name{simulate_model_pp}
\alias{simulate_model_pp}
\title{Simulate a Point Pattern from a scampr model}
\usage{
simulate_model_pp(
  model,
  domain.data,
  type = c("link", "response"),
  dens = c("posterior", "prior"),
  rseed = NA,
  expected.or.sampled.intensity = c("expected", "sampled")
)
}
\arguments{
\item{model}{a scampr model}

\item{domain.data}{optionally, data frame of point locations that adequately cover the domain of interest, as well as, predictors involved in the model. If missing model quadrature is used.}

\item{type}{a character string indicating the type of prediction to be returned. One of 'link' or 'response' for log-intensity or intensity respectively.}

\item{dens}{a character string indicating the probability density of the random effects to draw from.}

\item{rseed}{an integer for setting the random seed.}

\item{expected.or.sampled.intensity}{a character string indicating whether to simulate from the expected intensity (with respect to 'type' dens') or a randomly sampled intensity (with respect to 'dens'). One of 'expected' or 'sampled' respectively. This will work in conjuction with 'type' (changes expectation) and 'dens' (N(0,prior_var) or N(posterior_mean, posterior_var)). Toggling this depends on the use of simulation, e.g. 'sample' may be relevant to temporal models or for simulating future scenarios.}
}
\value{
point pattern object of class 'ppp' from spatstat
}
\description{
Simulate a Point Pattern from a scampr model
}
\examples{
dat <- scampr::gorillas
dat$elev <- scale(dat$elevation)
mod <- po(pres ~ elev, dat, model.type = "ipp")
pp <- simulate_model_pp(mod)
}
