% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.kfold.cv2.R,
%   R/spatial.kfold.cv_parallel.R
\name{spatial.kfold.cv_parallel}
\alias{spatial.kfold.cv_parallel}
\title{Performs a spatial hold-one-out K fold cross validation on a scampr model}
\usage{
spatial.kfold.cv_parallel(
  object,
  po.fold.id,
  pa.fold.id,
  trunc.pa.prob = 1e-07
)

spatial.kfold.cv_parallel(
  object,
  po.fold.id,
  pa.fold.id,
  trunc.pa.prob = 1e-07
)
}
\arguments{
\item{object}{a scampr model object}

\item{po.fold.id}{An integer or factor vector the same length as the presence-only data used in the model \code{object} - describes the CV fold that each location falls into.}

\item{pa.fold.id}{An integer or factor vector the same length as the presence/absence data used in the model \code{object} - describes the CV fold that each location falls into.}

\item{trunc.pa.prob}{A small positive number by which the predicted probability of presence is truncated. This can be used to ensure infinite values are avoided within the cross-validation.}
}
\value{
a list including- 'predicted.cll.pa': the predicted logLikelihood, conditional on the model predictors and latent field, on the held out presence/absence data. 'predicted.cll.po': the predicted logLikelihood, conditional on the model predictors and latent field, on the held out presence-only data. 'auc': the area under the ROC curve on the held out presence/absence sites. 'model': the model upon which the cross validation was performed.

a list including- 'predicted.cll.pa': the predicted logLikelihood, conditional on the model predictors and latent field, on the held out presence/absence data. 'predicted.cll.po': the predicted logLikelihood, conditional on the model predictors and latent field, on the held out presence-only data. 'auc': the area under the ROC curve on the held out presence/absence sites. 'model': the model upon which the cross validation was performed.
}
\description{
Performs a spatial hold-one-out K fold cross validation on a scampr model

Performs a spatial hold-one-out K fold cross validation on a scampr model (using parallelisation via \code{foreach})
}
\examples{
#' # Get the Eucalypt data
dat_po <- eucalypt[["po"]]
dat_pa <- eucalypt[["pa"]]

\dontrun{
# Fit with a shared latent field
res <- simple_basis_search_popa(pres ~ TMP_MIN + D_MAIN_RDS, Y ~ TMP_MIN,
po.data = dat_po, pa.data = dat_pa)
}
#' # Get the Eucalypt data
dat_po <- eucalypt[["po"]]
dat_pa <- eucalypt[["pa"]]

\dontrun{
# Fit with a shared latent field
res <- simple_basis_search_popa(pres ~ TMP_MIN + D_MAIN_RDS, Y ~ TMP_MIN,
po.data = dat_po, pa.data = dat_pa)
}
}
