% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scampr.R
\name{summary.scampr}
\alias{summary.scampr}
\title{Summary for objects of class 'scampr'}
\usage{
\method{summary}{scampr}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Summary for objects of class 'scampr'
}
