% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.starting.parameters.R
\name{update.starting.parameters}
\alias{update.starting.parameters}
\title{Internal scampr function that gets warm starting parameters to pass to TMB}
\usage{
\method{update}{starting.parameters}(
  new.start.pars,
  old.start.pars,
  target.model.type = c("laplace", "variational", "ipp")
)
}
\arguments{
\item{new.start.pars}{either a fitted model or named list of starting parameters}

\item{old.start.pars}{a named list of starting parameters}

\item{target.model.type}{a character string, one of 'laplace', 'variational' or 'ipp'.}
}
\value{
a data.frame (sparse or dense depending on parameter bf.matrix.type)
}
\description{
Internal scampr function that gets warm starting parameters to pass to TMB
}
\examples{
# Get the gorilla nesting data
dat <- gorillas
}
