% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.scampr.R
\name{vcov.scampr}
\alias{vcov.scampr}
\title{vcov for objects of class 'scampr'}
\usage{
\method{vcov}{scampr}(object, ..., getJointPrecision = T)
}
\arguments{
\item{object}{a scampr model object}

\item{...}{NA}

\item{getJointPrecision}{a logical indicating whether to calculate the full joint precision matrix, including random effects (TRUE) or just the fixed effects (FALSE).}
}
\value{
A matrix of the estimated covariances between the parameter estimates in the linear or non-linear predictor of the model. This should have row and column names corresponding to the parameter names given by the coef method.
}
\description{
vcov for objects of class 'scampr'
}
\examples{
# Get the gorilla nesting data
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Fit a scampr model to the point pattern
m <- scampr(pres ~ elev.std, data = dat, include.sre = F)

vcov(m)
}
