% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2im.R
\name{vec2im}
\alias{vec2im}
\title{Internal scampr function that converts a vector to spatstat image given spatial locations.}
\usage{
vec2im(vec, x.loc, y.loc)
}
\arguments{
\item{vec}{A vector of field values of equal length to locations x.loc and y.loc}

\item{x.loc}{A vector of horizontal locations in 2D}

\item{y.loc}{A vector of vertical locations in 2D}
}
\value{
a spatstat image (im) object
}
\description{
Internal scampr function that converts a vector to spatstat image given spatial locations.
}
\examples{
#' # Get the quadrature from the gorillas data
quad <- gorillas[gorillas$pres == 0, ]

elevation <- scampr:::vec2im(vec = quad$elevation, x.loc = quad$x, y.loc = quad$y)
}
