% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2mat.R
\name{vec2mat}
\alias{vec2mat}
\title{Title}
\usage{
vec2mat(vec, x.loc, y.loc)
}
\arguments{
\item{vec}{A vector of field values of equal length to locations x.loc and y.loc}

\item{x.loc}{A vector of horizontal locations in 2D}

\item{y.loc}{A vector of vertical locations in 2D}
}
\value{

}
\description{
Title
}
