% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.scampr.R
\name{AIC.scampr}
\alias{AIC.scampr}
\title{AIC for objects of class 'scampr'}
\usage{
\method{AIC}{scampr}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
AIC for objects of class 'scampr'
}
