% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.bf.matrix.R
\name{get.bf.matrix}
\alias{get.bf.matrix}
\title{Internal scampr function that calculates the basis function evaluations at point locations provided. From either a fitted model or data.frame of class 'bf.df'}
\usage{
get.bf.matrix(object, point.locations, bf.matrix.type = c("sparse", "dense"))
}
\arguments{
\item{object}{either a fitted model or data.frame of class 'bf.df' created by 'simple_basis'}

\item{point.locations}{either a matrix or data.frame describing the point locations in the same order as in 'object'}

\item{bf.matrix.type}{a character string, one of 'sparse' or 'dense' indicating whether to create sparse or dense matrix.}
}
\value{
a data.frame (sparse or dense depending on parameter bf.matrix.type)
}
\description{
Internal scampr function that calculates the basis function evaluations at point locations provided. From either a fitted model or data.frame of class 'bf.df'
}
\examples{
# Get the gorilla nesting data
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Set up a simple 2D grid of basis functions to fit a LGCP model to the data
bfs <- simple_basis(nodes.on.long.edge = 9, data = dat)

# Fit a LGCP model using variational approximation
m.lgcp_va <- scampr(pres ~ elev.std, data = dat, model.type = "PO", basis.functions = bfs)

# Get some new locations
new.sites <- dat[sample(1:nrow(dat), 10), c("x", "y")]

# Calculate the basis function matrix at the new sites
scampr:::get.bf.matrix(m.lgcp_va, new.sites)
}
