% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.TMB.data.input.R
\name{get.TMB.data.input}
\alias{get.TMB.data.input}
\title{Internal scampr function that creates a list of data and starting parameters for scampr models}
\usage{
get.TMB.data.input(
  formula,
  data,
  bias.formula,
  IDM.presence.absence.df,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  model.type = c("variational", "laplace", "ipp"),
  data.type = c("PO", "PA", "IDM"),
  basis.functions,
  bf.matrix.type = c("sparse", "dense"),
  starting.pars,
  po.biasing.basis.functions
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the fixed effects of the model to be fitted. The 'response' must be a binary that indicates whether a datum is a presence or: quadrature point (for point process models)/ absence (for binary models). See GLM function for further formula details.}

\item{data}{a data frame containing response and predictors within \code{formula}.}

\item{bias.formula}{an object of class "formula" (or one that can be coerced to that class) OR the character string "latent". In the formula case, this is a symbolic description of the predictors included to account for bias in the presence-only data (no response term is needed). In the case of fitting an integrated data model, \code{bias.formula = "latent"} will fit an approximate latent Gaussian field to account for the bias.}

\item{IDM.presence.absence.df}{an optional data frame. When fitting an integrated data model use this to pass in the presence/absence data.}

\item{coord.names}{a vector of character strings describing the column names of the coordinates in both data frames.}

\item{quad.weights.name}{a charater string of the column name of quadrature weights in the data.}

\item{model.type}{a character string indicating the type of model to be used. May be one of 'laplace' or 'variational' for Cox Processes involving spatially correlated errors or 'ipp' for a model that follows an inhomgeneous Poisson process.}

\item{data.type}{a character string indicating the type of data to be used. May be one of 'PO' (for a presence-only PPM) or 'PA' (for a presence/absence Binary GLM) or 'IDM' (for an integrated data model).}

\item{basis.functions}{an optional object of class 'Basis' created by \code{FRK::auto_basis()} or 'bf.df' created by \code{scampr::simple_basis()}. Either object describes a set of basis functions for approximating the latent Gaussian field. If NULL the model will use default \code{FRK::auto_basis()} with \code{max_basis = 0.25 * # of points}.}

\item{bf.matrix.type}{a character string, one of 'sparse' or 'dense' indicating whether to use sparse or dense matrix computations for the basis functions created.}

\item{starting.pars}{an optional named list or scampr model object that gives warm starting values for the parameters of the model.}

\item{po.biasing.basis.functions}{an optional extra set of basis functions that can be used when \code{bias.formula = "latent"}, otherwise \code{basis.functions} are used.}
}
\value{
list of elements required for TMB::MakeADFun
}
\description{
Internal scampr function that creates a list of data and starting parameters for scampr models
}
\examples{
# Get the flora data for one of the species
dat_po <- flora$po$sp1
dat_pa <- flora$pa

# Attach the quadrature to the PO data
dat_po <- rbind.data.frame(dat_po, flora$quad)

# Get the TMB data lists for a combined data model without latent field
tmb.input <- scampr:::get.TMB.data.input(pres ~ MNT + D.Main, sp1 ~ MNT, po.data = dat_po, pa.data = dat_pa, model.type = "ipp")
str(tmp.input)
}
