% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.scampr.R
\name{image.scampr}
\alias{image.scampr}
\title{Plot the image of a field over a model quadrature}
\usage{
\method{image}{scampr}(x, z, residual.type, ...)
}
\arguments{
\item{x}{A scampr model object}

\item{z}{Either a single characater string of the variable name (in the model data) or one of 'fitted', 'residuals'. Alternatively, a vector of numeric values to be plotted}

\item{residual.type}{an optional character string for residual type if z == 'residual'}

\item{...}{}
}
\value{

}
\description{
Plot the image of a field over a model quadrature
}
\examples{
# Get the Eucalypt data
dat_po <- eucalypt[["po"]]
dat_pa <- eucalypt[["pa"]]

# Fit an IPP model to the point pattern
m.ipp <- ippm(pres ~ TMP_MIN + D_MAIN_RDS, data = dat_po)

# Fit a combined data model
m.popa <- popa(pres ~ TMP_MIN + D_MAIN_RDS, Y ~ TMP_MIN, po.data = dat_po, pa.data = dat_pa, model.type = "ipp")

# Fit presence/absence model
m.pa <- pa(Y ~ TMP_MIN, pa.data = dat_pa, model.type = "ipp")

image(m.ipp, "fitted")
image(m.popa, "fitted")
image(m.pa, "fitted")
image(m.ipp, "TMP_MIN")
image(m.popa, "residuals")
}
