% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.scampr.R
\name{image.scampr}
\alias{image.scampr}
\title{Plot the image of a field over a model quadrature}
\usage{
\method{image}{scampr}(x, z, ...)
}
\arguments{
\item{x}{A scampr model object}

\item{z}{A vector of length equal to the number (and in order) of the quadrature in model x}

\item{...}{}
}
\value{

}
\description{
Plot the image of a field over a model quadrature
}
