% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.scampr.R
\name{image.scampr}
\alias{image.scampr}
\title{Plot the image of a field over a model quadrature}
\usage{
\method{image}{scampr}(x, z, residual.type, ...)
}
\arguments{
\item{x}{A scampr model object}

\item{z}{Either a single characater string of the variable name (in the model data) or one of 'fitted', 'residuals'. Alternatively, a vector of numeric values to be plotted}

\item{residual.type}{an optional character string for residual type if z == 'residual'}

\item{...}{NA}
}
\value{
See fields::image.plot()
}
\description{
An image plotting function that uses fields::image.plot() to display the specified variable as a field over the quadrature.
}
\examples{
# Fit a combined data model
m.popa <- popa(pres ~ TMP_MIN + D_MAIN_RDS, Y ~ TMP_MIN,
po.data = eucalypt[["po"]], pa.data = eucalypt[["pa"]],
model.type = "ipp")

image(m.popa, "fitted")
\dontrun{
image(m.popa, "TMP_MIN")
image(m.popa, "residuals")
}
}
