% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.scampr.R
\name{image.scampr}
\alias{image.scampr}
\title{Plot the image of a field over a model quadrature ### WIP ###}
\usage{
\method{image}{scampr}(x, z, domain.data, residual.type, residual.smoothing = 0.5, ...)
}
\arguments{
\item{x}{a scampr model object}

\item{z}{Either a single characater string of the variable name (in the model data) or one of 'fitted', 'residuals'. Alternatively, a vector of numeric values to be plotted}

\item{domain.data}{a data frame containing (at least) the horizontal and vertical coordinates as used in the model 'x' as well as the variable name of 'z' if specified. If missing, the quadrature points of the model will be used.}

\item{residual.type}{an optional character string for residual type if z == 'residual'}

\item{residual.smoothing}{an optional numeric for the scale of residual smoothing (theta in fields::image.smooth())}

\item{...}{additional plotting arguments}
}
\value{
See fields::image.plot()
}
\description{
An image plotting function that uses fields::image.plot() to display the specified variable as a field over the quadrature.
}
\examples{
# Get the flora data for one of the species
dat_po <- flora$po$sp1
dat_pa <- flora$pa

# Attach the quadrature to the PO data
dat_po <- rbind.data.frame(dat_po, flora$quad)

# Fit a combined data model
m.comb <- scampr(pres ~ MNT + D.Main, dat_po, sp1 ~ MNT,
dat_pa, model.type = "ipp")

image(m.comb, "fitted")
\dontrun{
image(m.comb, "MNT")
image(m.comb, "residuals")
}
}
