% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ippm.R
\name{ippm}
\alias{ippm}
\title{Inhomogeneous Poisson Process Model for Point Patterns}
\usage{
ippm(
  formula,
  data,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  starting.pars,
  se = TRUE
)
}
\arguments{
\item{formula}{formula describing fixed effects of the linear predictor. Response must be the presence/quadrature identifier.}

\item{data}{data frame containing predictors at both presence-records and quadrature}

\item{coord.names}{vector of character strings describing the column names of the coordinates in data}

\item{quad.weights.name}{charater string of the column name of quadrature weights in data}

\item{starting.pars}{optional named list or scampr model that gives warm starts for the parameters of the model}

\item{se}{logical indicating whether standard errors should be calculated}
}
\value{
scampr model object
}
\description{
Fit an inhomogeneous Poisson process model to presence records using numerical quadrature
}
\examples{
#' # Get the gorilla nesting data
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Fit an IPP model to the point pattern
m.ipp <- ippm(pres ~ elev.std, data = dat)
}
