% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpm.R
\name{lgcpm}
\alias{lgcpm}
\title{Approximate log-Gaussian Cox Process Model for a Point Pattern}
\usage{
lgcpm(
  formula,
  data,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  FRK.basis.functions,
  simple.basis,
  approx.with = c("laplace", "variational"),
  se = TRUE,
  starting.pars,
  subset,
  na.action
)
}
\arguments{
\item{formula}{formula describing fixed effects of the linear predictor. Response must be the presence/quadrature binary (1/0) column name.}

\item{data}{data frame containing predictors at both presence-records and quadrature}

\item{coord.names}{vector of character strings describing the column names of the coordinates in data}

\item{quad.weights.name}{charater string of the column name of quadrature weights in data}

\item{FRK.basis.functions}{object of class 'Basis' from FRK package. If neither 'FRK.basis.functions' nor 'simple.basis' is specified will use default FRK::auto_basis with 2 spatial resolutions}

\item{simple.basis}{Alternative to 'FRK.basis.functions', a data.frame of basis functions information created by 'simple_basis()'.}

\item{approx.with}{charater indicating the type of approximation to use for the intractible marginalisation: variational or Laplace}

\item{se}{logical indicating whether standard errors should be calculated}

\item{starting.pars}{optional named list or scampr model that gives warm starts for the parameters of the model}

\item{subset}{optional subset of the data}

\item{na.action}{optional way of handling NA's in the data, default is omit}
}
\value{
scampr model object
}
\description{
Blah blah blah
}
