% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FRK_bf_matrix.R
\name{make_FRK_bf_matrix}
\alias{make_FRK_bf_matrix}
\title{Create an FRK basis function matrix}
\usage{
make_FRK_bf_matrix(
  data,
  coord.names = c("x", "y"),
  manifold = FRK::plane(),
  regular = 1,
  nres = 2,
  prune = 0,
  max_basis = NULL,
  subsamp = 10000,
  type = c("bisquare", "Gaussian", "exp", "Matern32"),
  isea3h_lo = 2,
  bndary = NULL,
  scale_aperture = ifelse(is(manifold, "sphere"), 1, 1.25),
  verbose = 0L,
  ...
)
}
\arguments{
\item{data}{data frame containing locations of both presence-records and quadrature}

\item{coord.names}{vector of character strings describing the column names of the coordinates in data}

\item{manifold}{see FRK package documentation}

\item{regular}{see FRK package documentation}

\item{nres}{see FRK package documentation}

\item{prune}{see FRK package documentation}

\item{max_basis}{see FRK package documentation}

\item{subsamp}{see FRK package documentation}

\item{type}{see FRK package documentation}

\item{isea3h_lo}{see FRK package documentation}

\item{bndary}{see FRK package documentation}

\item{scale_aperture}{see FRK package documentation}

\item{verbose}{see FRK package documentation}

\item{...}{}
}
\value{
matrix describing the basis functions (columns) value at each location in the data (rows)
}
\description{
Essentially a wrapper on the FRK package function to make correct format
}
