% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scampr.R
\name{plot.scampr}
\alias{plot.scampr}
\title{Plot function for objects of class 'scampr'}
\usage{
plot.scampr(x, add.points = F)
}
\arguments{
\item{x}{a scampr model object}

\item{add.points}{logical indicating whether to add the presence point locations to plots}

\item{...}{}
}
\value{

}
\description{
Plot function for objects of class 'scampr'
}
\examples{
# Get the Eucalypt data
dat_po <- eucalypt[["po"]]
dat_pa <- eucalypt[["pa"]]

# Fit an IPP model to the point pattern
m.ipp <- ippm(pres ~ elev.std, data = dat)

# Fit a combined data model
m.popa <- popa(pres ~ TMP_MIN + D_MAIN_RDS, Y ~ TMP_MIN, po.data = dat_po, pa.data = dat_pa, model.type = "ipp")

# Fit presence/absence model
m.pa <- pa(Y ~ TMP_MIN, pa.data = dat_pa, model.type = "ipp")

plot(m.ipp)
plot(m.popa)
plot(m.pa)
}
