% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scampr.R
\name{plot.scampr}
\alias{plot.scampr}
\title{Plot function for objects of class 'scampr'}
\usage{
plot.scampr(x, add.points = F)
}
\arguments{
\item{x}{a scampr model object}

\item{add.points}{logical indicating whether to add the presence point locations to plots}

\item{...}{}
}
\value{

}
\description{
Plot function for objects of class 'scampr'
}
