% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scampr.R
\name{plot.scampr}
\alias{plot.scampr}
\title{Plot diagnositcs for objects of class 'scampr' ### WIP ###}
\usage{
\method{plot}{scampr}(x, ..., which = c("residuals", "fitted"), add.points = F)
}
\arguments{
\item{x}{a scampr model object}

\item{...}{additional plotting arguments}

\item{which}{if a subset of the plots is required, specify one of 'residuals' or 'fitted': plots Pearson residuals and fitted intensity respectively for PPM (for a PA data model: plots fitted vs. raw residuals and the survey data for 'residuals' and 'fitted' resp.).}

\item{add.points}{logical indicating whether to add the presence point locations to plots}
}
\value{
See spatstat::plot.ppm
}
\description{
Currently plots residuals as in 'spatstat', as well as the fitted intensity, over the quadrature used to model the point pattern. In the case of presence/absence data fitted vs. residuals is plotted.
}
\examples{
# Get the flora data for one of the species
dat_po <- flora$po$sp1
dat_pa <- flora$pa

# Attach the quadrature to the PO data
dat_po <- rbind.data.frame(dat_po, flora$quad)

# Fit an IPP model to the point pattern
m.ipp <- scampr(pres ~ MNT + D.Main, data = dat_po, model.type = "ipp")

# Fit a combined data model
m.comb <- scampr(pres ~ MNT + D.Main, dat_po, sp1 ~ MNT,
dat_pa, model.type = "ipp")

# Fit presence/absence model
m.pa <- scampr(pa.formula = sp1 ~ MNT, pa.data = dat_pa, model.type = "ipp")

plot(m.ipp)
plot(m.comb)
plot(m.pa)
}
