% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_search.R
\name{plot_search}
\alias{plot_search}
\title{Plot basis function search results - for objects of class 'basis.search' ### WIP ###}
\usage{
plot_search(x, ...)
}
\arguments{
\item{x}{an object of class 'basis.search'}

\item{...}{additional plotting arguments}
}
\description{
Plots the results of a simple basis function search created by \code{scampr::basis.search()}.
}
\examples{
# Get the flora data for one of the species
dat_po <- flora$po$sp1
dat_pa <- flora$pa

# Create the basis functions on the flora quadrature
bfs <-  simple_basis(9, flora$quad)

# Plot the bi-square basis functions
plot(bfs)

# Fit a combined data model
m.comb <- scampr(pres ~ MNT + D.Main, dat_po, sp1 ~ MNT,
dat_pa, model.type = "ipp")

# Fit presence/absence model
m.pa <- scampr(pa.formula = sp1 ~ MNT, pa.data = dat_pa, model.type = "ipp")

plot(m.ipp)
plot(m.comb)
plot(m.pa)
}
