% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_search.R
\name{plot_search}
\alias{plot_search}
\title{Plot basis function search results - for objects of class 'basis.search' ### WIP ###}
\usage{
plot_search(x, ..., plot.metric = 1:6)
}
\arguments{
\item{x}{an object of class 'basis.search'}

\item{...}{additional plotting arguments}

\item{plot.metric}{one or more numbers from 1 to 6 indicating the metric(s) to be plotted. 1: Fitted approximate log-likelihood. 2: Fitted approx. log-Likelihood (Presence-only component). 3: Fitted approx. log-Likelihood (Presence/absence component). 4: Predicted Conditional log-Likelihood (Presence-only Data). 5: Predicted Conditional log-Likelihood (Presence/absence Data). 6: Area Under the Operator/Receiver Characteristic Curve (AUC) on the Presence/absence data.}
}
\description{
Plots the results of a simple basis function search created by \code{scampr::basis.search()}.
}
\examples{
# Get the flora data for one of the species
dat_po <- flora$po$sp1
dat_pa <- flora$pa

# Create the basis functions on the flora quadrature
bfs <-  simple_basis(9, flora$quad)

# Plot the bi-square basis functions
plot(bfs)

# Fit a combined data model
m.comb <- scampr(pres ~ MNT + D.Main, dat_po, sp1 ~ MNT,
dat_pa, model.type = "ipp")

# Fit presence/absence model
m.pa <- scampr(pa.formula = sp1 ~ MNT, pa.data = dat_pa, model.type = "ipp")

plot(m.ipp)
plot(m.comb)
plot(m.pa)
}
