% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.R
\name{po}
\alias{po}
\title{Point Process Model for Presence-only Data}
\usage{
po(
  po.formula,
  po.data,
  coord.names = c("x", "y"),
  quad.weights.name = "quad.size",
  FRK.basis.functions,
  simple.basis,
  model.type = c("variational", "laplace", "ipp"),
  bf.matrix.type = c("sparse", "dense"),
  se = TRUE,
  starting.pars,
  subset
)
}
\arguments{
\item{po.formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the data model to be fitted. The 'response' must be a binary that indicates whether a datum is a presence or quadrature point. See GLM function for further formula details.}

\item{po.data}{a data frame containing predictors at both presence-records and quadrature as well as the po.formula 'response'.}

\item{coord.names}{a vector of character strings describing the column names of the coordinates in both data frames}

\item{quad.weights.name}{a charater string of the column name of quadrature weights in the po.data}

\item{FRK.basis.functions}{an optional object of class 'Basis' from FRK package. If neither 'FRK.basis.functions' nor 'simple.basis' is specified will use default FRK::auto_basis with 2 spatial resolutions.}

\item{simple.basis}{an alternative to 'FRK.basis.functions': a data.frame of basis functions information created by 'simple_basis()'.}

\item{model.type}{a character string indicating the type of model to be used. May be one of 'laplace' or 'variational' for Cox Processes involving spatially correlated errors or 'ipp' for a model that follows an inhomgeneous Poisson process.}

\item{bf.matrix.type}{a character string, one of 'sparse' or 'dense' indicating whether to use sparse or dense matrix computations for the basis functions created.}

\item{se}{a logical indicating whether standard errors should be calculated.}

\item{starting.pars}{an optional named list or scampr model object that gives warm starting values for the parameters of the model.}

\item{subset}{an optional vector describing a subset of the data to be used.}
}
\value{
a scampr model object
}
\description{
Fit either an inhomogeneous Poisson, or log-Gaussian Cox, process model to presence records. This uses numerical quadrature  (provided with the data, see e.g. scampr::gorillas) to approx. the spatial integral. If fitting a LGCP, uses one of either a Laplace or variational approximation to marginalise over the latent field.
}
\examples{
# Get the gorilla nesting data
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Fit an IPP model to the point pattern
m.ipp <- po(pres ~ elev.std, po.data = dat, model.type = "ipp")

# Set up a simple 2D grid of basis functions to fit a LGCP model to the data
bfs <- simple_basis(nodes.on.long.edge = 9, data = dat)

# Fit a LGCP model using variational approximation
m.lgcp_va <- po(pres ~ elev.std, po.data = dat, model.type = "variational", simple.basis = bfs)

\dontrun{
# Fit a LGCP model using Laplace approximation
m.lgcp_lp <- po(pres ~ elev.std, po.data = dat, model.type = "laplace", simple.basis = bfs)
}
}
