% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scampr.R
\name{print.scampr}
\alias{print.scampr}
\title{Print for objects of class 'scampr'}
\usage{
\method{print}{scampr}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Print for objects of class 'scampr'
}
