% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.scampr.R
\name{residuals.scampr}
\alias{residuals.scampr}
\title{Extract Model Residuals for objects of class 'scampr'}
\usage{
\method{residuals}{scampr}(object, type = "raw")
}
\arguments{
\item{type}{As in residuals.ppm: A string indicating the type of residuals or weights to be used. Current options are "raw" for the raw residuals, "inverse" for the inverse-lambda residuals, and "pearson" for the Pearson residuals.}
}
\value{

}
\description{
Extract Model Residuals for objects of class 'scampr'
}
