% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.scampr.R
\name{residuals.scampr}
\alias{residuals.scampr}
\title{Extract Model Residuals for objects of class 'scampr'}
\usage{
\method{residuals}{scampr}(
  object,
  ...,
  type = c("raw", "inverse", "pearson"),
  data.type = c("po", "pa")
)
}
\arguments{
\item{object}{a scampr model object}

\item{...}{NA}

\item{type}{As in residuals.ppm: A string indicating the type of residuals or weights to be used. Current options are "raw" for the raw residuals, "inverse" for the inverse-lambda residuals, and "pearson" for the Pearson residuals.}

\item{data.type}{a character string indicating which data type to produce residuals from.}
}
\value{
a numeric vector, of length of the model data, containing the extracted residuals.
}
\description{
Extract Model Residuals for objects of class 'scampr'
}
\examples{
# Get the flora data for one of the species
dat_po <- flora$po$sp1
dat_pa <- flora$pa

# Attach the quadrature to the PO data
dat_po <- rbind.data.frame(dat_po, flora$quad)

# Set up a simple 2D grid of basis functions to fit a LGCP model to the data
bfs <- simple_basis(nodes.on.long.edge = 9, data = dat_po)

# Fit an IPP model to the point pattern
m.ipp <- scampr(pres ~ MNT + D.Main, dat_po, model.type = "ipp")

# Fit a combined data model
m.comb <- scampr(pres ~ MNT + D.Main, dat_po, sp1 ~ MNT,
dat_pa, model.type = "ipp")

# Fit presence/absence model
m.pa <- scampr(pa.formula = sp1 ~ MNT, pa.data = dat_pa, model.type = "ipp")

\dontrun{
# Fit a LGCP model to the point pattern
m.lgcp_va <- scampr(pres ~ MNT + D.Main, dat_po,
model.type = "variational", simple.basis = bfs)
# Or
m.lgcp_lp <- scampr(pres ~ MNT + D.Main, dat_po,
approx.with = "laplace", simple.basis = bfs)
}
residuals(m.ipp)
residuals(m.comb)
residuals(m.comb, data.type = "pa")
residuals(m.pa, data.type = "pa")
\dontrun{
residuals(m.lgcp_va)
residuals(m.lgcp_lp, type = "pearson")
}
}
