% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_basis.R
\name{simple_basis}
\alias{simple_basis}
\title{Create a single resolution, regular grid of bisquare basis function nodes in flat 2D space}
\usage{
simple_basis(nodes.on.long.edge, data, coord.names = c("x", "y"))
}
\arguments{
\item{nodes.on.long.edge}{integer describing the number of basis function nodes to place along the longest edge of the domain}

\item{data}{a data frame containing the two coordinates described by 'coord.names'}

\item{coord.names}{vector of character strings describing the names of the coordinates in 'data'. Ordered horizontal axis to vertical}
}
\value{

}
\description{
Create a single resolution, regular grid of bisquare basis function nodes in flat 2D space
}
