% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lgcp_pp.R
\name{simulate_lgcp_pp}
\alias{simulate_lgcp_pp}
\title{Simulate a point pattern from log-Gaussian Cox Process}
\usage{
simulate_lgcp_pp(
  Intercept = -3,
  Slope = 1.25,
  latent.variance = 1,
  latent.range = 15,
  rseed = NA
)
}
\arguments{
\item{Intercept}{Fixed effect intercept - used to control the expected number of points}

\item{Slope}{Fixed effect slope - controls the effect of the spatial covariate on the point pattern}

\item{latent.variance}{The marginal variance of the latent Gaussian field}

\item{latent.range}{The correlation range of the latent Gaussian field}

\item{rseed}{integer for setting the random seed}
}
\value{
list of simulated data frame, sp point pattern and information on the attributes of the simulation
}
\description{
Simulate's a point pattern from a LGCP with an intercept and slope on a single spatial covariate. Control the correlation range and marginal variance of the latent field
}
