% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_po_pa.R
\name{simulate_po_pa}
\alias{simulate_po_pa}
\title{Simulate presence-only data (point pattern from log-Gaussian Cox Process) and presence/absence data (from the same underlying process)}
\usage{
simulate_po_pa(
  Intercept.env = -1.75,
  Intercept.bias = -2,
  Beta = c(-1.2, 0.75),
  Tau = c(1.3, -0.8),
  latent.variance = 1,
  latent.range = 30,
  sites.sampled = 1000,
  rseed = NA
)
}
\arguments{
\item{Intercept.env}{Fixed effect intercept}

\item{Intercept.bias}{Bias effect intercept}

\item{Beta}{length 2 vector of env. effects}

\item{Tau}{length 2 vector of bias effects}

\item{latent.variance}{variance for the latent field}

\item{latent.range}{correlation range for the latent field}

\item{sites.sampled}{number of survey sites to be sampled}

\item{rseed}{integer for setting the random seed}
}
\value{

}
\description{
Simulate presence-only data (point pattern from log-Gaussian Cox Process) and presence/absence data (from the same underlying process)
}
