% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scampr.R
\name{summary.scampr}
\alias{summary.scampr}
\title{Summary for objects of class 'scampr'}
\usage{
\method{summary}{scampr}(object, ...)
}
\arguments{
\item{object}{A scampr model object}

\item{...}{NA}
}
\value{
printed argument displaying summary information of the model fit.
}
\description{
Summary for objects of class 'scampr'
}
\examples{
# Get the gorilla nesting data
data(gorillas, package = "scampr")
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Fit a scampr model to the point pattern
m <- ipp(pres ~ elev.std, data = dat)

summary(m)
}
