% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.starting.parameters.R
\name{update.starting.parameters}
\alias{update.starting.parameters}
\title{Internal scampr function that gets warm starting parameters to pass to TMB}
\usage{
\method{update}{starting.parameters}(
  new.start.pars,
  old.start.pars,
  target.approx.type = c("laplace", "variational", "not_sre")
)
}
\arguments{
\item{new.start.pars}{either a fitted model or named list of starting parameters}

\item{old.start.pars}{a named list of starting parameters}

\item{target.approx.type}{a character string, one of 'laplace', 'variational' or 'not_sre'.}
}
\value{
a data.frame (sparse or dense depending on parameter bf.matrix.type)
}
\description{
Internal scampr function that gets warm starting parameters to pass to TMB
}
\examples{
# Get the gorilla nesting data
data(gorillas, package = "scampr")
dat <- gorillas

# Standardise the elevation covariate
dat$elev.std <- scale(dat$elevation)

# Fit a scampr model to the point pattern
m <- scampr(pres ~ elev.std, data = dat, include.sre = F)

# Get new start parameters
update.startingp.parameters(m, list(fixed = rep(0, ncol(dat.list$X_PO_pres)),
  random = rep(0, ncol(dat.list$Z_PO_pres)),
  log_variance_component = var.starts))
}
