% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.scampr.R
\name{vcov.scampr}
\alias{vcov.scampr}
\title{vcov for objects of class 'scampr'}
\usage{
\method{vcov}{scampr}(object)
}
\arguments{
\item{object}{A scampr model object}
}
\value{

}
\description{
vcov for objects of class 'scampr'
}
